+++ include/JSystem/JKernel/JKRDvdRipper.h
@@ -28,10 +28,10 @@ struct JKRDvdRipper {
 	};
 
 	static void* loadToMainRAM(const char*, u8*, JKRExpandSwitch, u32, JKRHeap*, EAllocDirection, u32, int*, u32*);
-	static void* loadToMainRAM(long, u8*, JKRExpandSwitch, u32, JKRHeap*, EAllocDirection, u32, int*, u32*);
+	static void* loadToMainRAM(s32, u8*, JKRExpandSwitch, u32, JKRHeap*, EAllocDirection, u32, int*, u32*);
 	static void* loadToMainRAM(JKRDvdFile*, u8*, JKRExpandSwitch, u32, JKRHeap*, EAllocDirection, u32, int*, u32*);
 	static void* loadToMainRAMAsync(const char*, u8*, JKRExpandSwitch, u32, JKRHeap*, u32*);
-	static void* loadToMainRAMAsync(long, u8*, JKRExpandSwitch, u32, JKRHeap*, u32*);
+	static void* loadToMainRAMAsync(s32, u8*, JKRExpandSwitch, u32, JKRHeap*, u32*);
 
 	static void syncAll(int);
 	static void countLeftSync();
