+++ src/Dolphin/MSL_C/MSL_Common/float.c
@@ -1,8 +1,10 @@
-long __float_nan[]     = { 0x7FFFFFFF };
-long __float_huge[]    = { 0x7F800000 };
-long __double_max[]    = { 0x7FEFFFFF, 0xFFFFFFFF };
-long __double_huge[]   = { 0x7FF00000, 0 };
-long __extended_min[]  = { 0x00100000, 0 };
-long __extended_max[]  = { 0x7FEFFFFF, 0xFFFFFFFF };
-long __float_max[]     = { 0x7F7FFFFF };
-long __float_epsilon[] = { 0x34000000 };
+#include "types.h"
+
+s32 __float_nan[]     = { 0x7FFFFFFF };
+s32 __float_huge[]    = { 0x7F800000 };
+s32 __double_max[]    = { 0x7FEFFFFF, 0xFFFFFFFF };
+s32 __double_huge[]   = { 0x7FF00000, 0 };
+s32 __extended_min[]  = { 0x00100000, 0 };
+s32 __extended_max[]  = { 0x7FEFFFFF, 0xFFFFFFFF };
+s32 __float_max[]     = { 0x7F7FFFFF };
+s32 __float_epsilon[] = { 0x34000000 };
