+++ src/Dolphin/MSL_C/MSL_Common/strtoul.c
@@ -23,14 +23,14 @@ enum scan_states {
  * @note Address: 0x800CBEE0
  * @note Size: 0x378
  */
-unsigned long __strtoul(int base, int max_width, int (*ReadProc)(void*, int, int), void* ReadProcArg, int* chars_scanned, int* negative,
+u32 __strtoul(int base, int max_width, int (*ReadProc)(void*, int, int), void* ReadProcArg, int* chars_scanned, int* negative,
                         int* overflow)
 {
 	int scan_state          = start;
 	int count               = 0;
 	int spaces              = 0;
-	unsigned long value     = 0;
-	unsigned long value_max = 0;
+	u32 value     = 0;
+	u32 value_max = 0;
 	int c;
 
 	*negative = *overflow = 0;
@@ -159,15 +159,15 @@ unsigned long __strtoul(int base, int max_width, int (*ReadProc)(void*, int, int
  * @note Address: 0x800CBAD4
  * @note Size: 0x40C
  */
-unsigned long long __strtoull(int base, int max_width, int (*ReadProc)(void*, int, int), void* ReadProcArg, int* chars_scanned,
+u64 __strtoull(int base, int max_width, int (*ReadProc)(void*, int, int), void* ReadProcArg, int* chars_scanned,
                               int* negative, int* overflow)
 {
 	int scan_state               = start;
 	int count                    = 0;
 	int spaces                   = 0;
-	unsigned long long value     = 0;
-	unsigned long long value_max = 0;
-	unsigned long long ullmax    = ULLONG_MAX;
+	u64 value     = 0;
+	u64 value_max = 0;
+	u64 ullmax    = ULLONG_MAX;
 	int c;
 
 	*negative = *overflow = 0;
@@ -295,9 +295,9 @@ unsigned long long __strtoull(int base, int max_width, int (*ReadProc)(void*, in
  * @note Address: 0x800CBA28
  * @note Size: 0xAC
  */
-unsigned long strtoul(const char* str, char** end, int base)
+u32 strtoul(const char* str, char** end, int base)
 {
-	unsigned long value;
+	u32 value;
 	int count, negative, overflow;
 
 	__InStrCtrl isc;
@@ -333,10 +333,10 @@ void strtoull(void)
  * @note Address: 0x800CB938
  * @note Size: 0xF0
  */
-long strtol(const char* str, char** end, int base)
+s32 strtol(const char* str, char** end, int base)
 {
-	unsigned long uvalue;
-	long svalue;
+	u32 uvalue;
+	s32 svalue;
 	int count, negative, overflow;
 
 	__InStrCtrl isc;
@@ -353,7 +353,7 @@ long strtol(const char* str, char** end, int base)
 		svalue = (negative ? -LONG_MIN : LONG_MAX);
 		errno  = ERANGE;
 	} else {
-		svalue = (negative ? (long)-uvalue : (long)uvalue);
+		svalue = (negative ? (s32)-uvalue : (s32)uvalue);
 	}
 
 	return svalue;
